NewsPageAdmin VerifyLink structure fix + some small fixes
